Move pubsub from sdk to builder
Pubsub is only used for builder/canvas communication
while sdk is responsible for rendering sites.
Here moved pubsub to builder. Now we can drop nanoevents
from cli dependencies.
