+++ apps/builder/app/shared/pubsub/index.ts
@@ -1,4 +1,4 @@
-import { createPubsub } from "@webstudio-is/react-sdk";
+import { createPubsub } from "./create";
 
 // this interface is meant to be augmented by the user
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
