+++ packages/react-sdk/package.json
@@ -41,7 +41,6 @@
     "@webstudio-is/sdk": "workspace:^",
     "html-tags": "^3.3.1",
     "jsep": "^1.3.8",
-    "nanoevents": "^8.0.0",
     "nanoid": "^4.0.2",
     "nanostores": "^0.9.3",
     "no-case": "^3.0.4",
