--- packages/react-sdk/src/pubsub/index.ts
@@ -1,1 +1,0 @@
-export * from "./create";
